Mise au propre des fonctions dans son fichier respectif
